Extract repositories to nest lib
