+++ libs/repository/src/index.ts
@@ -1,0 +1,2 @@
+export * from './repository.module';
+export * from './services';
