+++ libs/repository/src/repository.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { UserRepository, PrismaService } from './services';
+
+@Module({
+  providers: [PrismaService, UserRepository],
+  exports: [UserRepository],
+})
+export class RepositoryModule {}
