+++ libs/repository/src/services/index.ts
@@ -1,0 +1,2 @@
+export * from './prisma.service';
+export * from './user.repository';
