+++ src/controllers/authentication/authentication.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { AuthenticationServiceModule } from '../domain/service/authentication/authentication.service.module';
 import { AuthenticationController } from './authentication.controller';
 import { SignUpCommandHandler } from './command/sign-up.command.handler';
+import { AuthenticationServiceModule } from '../../domain/service/authentication/authentication.service.module';
 
 const commandHandlers = [SignUpCommandHandler];
