+++ src/domain/service/user/user.service.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
+import { RepositoryModule } from '@mp/repository';
 import { UserService } from './user.service';
-import { UserRepository } from '../../../libs/repository/user.repository';
-import { PrismaService } from '../../../prisma.service';
 
 @Module({
-  providers: [UserService, UserRepository, PrismaService],
+  imports: [RepositoryModule],
+  providers: [UserService],
   exports: [UserService],
 })
 export class UserServiceModule {}
