+++ src/domain/service/user/user.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { User } from '../../entity/user.entity';
-import { UserRepository } from '../../../libs/repository/user.repository';
 import { UserCreationDto } from '../../../controllers/authentication/dto/user-creation.dto';
+import { UserRepository } from '@mp/repository';
+
 
 @Injectable()
 export class UserService {
