+++ test/jest-e2e.json
@@ -1,9 +1,17 @@
 {
-  "moduleFileExtensions": ["js", "json", "ts"],
+  "moduleFileExtensions": [
+    "js",
+    "json",
+    "ts"
+  ],
   "rootDir": ".",
   "testEnvironment": "node",
   "testRegex": ".e2e-spec.ts$",
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
+  },
+  "moduleNameMapper": {
+    "@mp/repository": "<rootDir>/../libs/repository/src",
+    "@mp/repository/(.*)": "<rootDir>/../libs/repository/src/$1"
   }
 }
