+++ tsconfig.json
@@ -17,6 +17,11 @@
     "forceConsistentCasingInFileNames": true,
     "noImplicitAny": true,
     "strictBindCallApply": true,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+    "paths": {
+      "@mp/repository": [
+        "libs/repository/src"
+      ]
+    }
   }
 }
